Add GAE and GKE fluentd Handlers
On GAE and GKE with the plugin installed,
there is a fluentd plugin that collects logs
from files. However without the right formatting,
metadata like log_level is lost. Furthermore,
the fluentd agents are configured to set the
correct resources types, which could be
done in the main handler as well, but it’s
easier to rely on the fluentd configurations.

This adds two new handlers and some helper
functions to detect when they should be used.